@@ -4,7 +4,7 @@
 		<OutputType>Exe</OutputType>
 		<PlatformTarget>AnyCPU</PlatformTarget>
 		<TieredCompilation>false</TieredCompilation>
-		<TargetFrameworks>netcoreapp3.0;netcoreapp2.1;net472</TargetFrameworks>
+		<TargetFrameworks>netcoreapp3.1;net472</TargetFrameworks>
 	</PropertyGroup>
 
 	<ItemGroup>
