@@ -5,7 +5,7 @@
 //	</auto-generated>
 //------------------------------------------------------------------------------
 
-#if HWINTRINSICS && !HWINTRINSICS_EXP
+#if HWINTRINSICS
 using System.Runtime.Intrinsics;
 using System.Runtime.Intrinsics.X86;
 using System.Runtime.InteropServices;
